feat(donations): add donation routes and integrate with chat service
- Add new donation routes to the main application router
- Move donation types to shared types file
- Make donation service async and integrate with chat service
- Implement singleton pattern with async initialization for donation service
